+++ src/Radi.js
@@ -14,10 +14,13 @@ import GLOBALS from './consts/GLOBALS';
 import { isWatchable, EMPTY_NODE } from './index';
 import EventService from './EventService';
 import PopulateService from './PopulateService';
+import Link from './Link';
 
 export default class Radi {
   constructor(o) {
     this.__path = 'this';
+
+    this.linkNum = 0;
 
     this.addNonEnumerableProperties({
       $mixins: o.$mixins,
@@ -36,9 +39,9 @@ export default class Radi {
       // TODO: REMOVE let _r = {r}; FIXME
       $view: new Function('r', 'list', 'll', 'cond', `let _r2 = { default: r }; return ${o.$view}`)(
         r.bind(this),
-        list.bind(this),
-        ll.bind(this),
-        cond.bind(this),
+        this.list,
+        this.ll,
+        this.cond,
       ),
     });
 
@@ -154,123 +157,68 @@ export default class Radi {
     this.mount();
     return this.$html;
   }
-}
 
+  cond(a, e) {
+    return new Condition(a, e, this);
+  }
 
-export function ll(fn, watch, c) {
-  return watch ? link(this, fn, watch, c.split(',')) : fn;
-}
+  list(data, act) {
+    if (!data) return '';
+    let link;
+    const fragment = document.createDocumentFragment();
+    const toplink = EMPTY_NODE.cloneNode();
 
-let linkNum = 0;
+    fragment.appendChild(toplink);
 
-export const link = (radiInstance, fn, watch, txt) => {
-  const args = {
-    s: null, a: [], t: [], f: fn.toString(),
+    const cache = data.source[data.prop] || [];
+    const cacheLen = cache.length || 0;
+
+    if (Array.isArray(cache)) {
+      for (let i = 0; i < cacheLen; i++) {
+        fragment.appendChild(act.call(this, cache[i], i));
+      }
+    } else {
+      let i = 0;
+      for (const key in cache) {
+        fragment.appendChild(act.call(this, cache[key], key, i));
+        i++;
+      }
+    }
+
+    link = fragment.lastChild;
+
+    const w = (a, b) => {
+      if (a === 0) return;
+      if (a > 0) {
+        const len = b.length;
+        const start = len - a;
+        for (let i = start; i < len; i++) {
+          fragment.appendChild(act.call(this, b[i], i));
+        }
+        const temp = fragment.lastChild;
+        link.parentElement.insertBefore(fragment, link.nextSibling);
+        link = temp;
+        return;
+      }
+      for (let i = 0; i < Math.abs(a); i++) {
+        const templink = link.previousSibling;
+        link.parentElement.removeChild(link);
+        link = templink;
+      }
+    };
+
+    if (cache.__path) {
+      let len = cacheLen;
+      this.$eventService.on(cache.__path, (e, v) => {
+        w(v.length - len, v);
+        len = v.length;
+      });
+    }
+
+    return fragment;
   };
 
-  if (
-    txt.length === 1 &&
-    fn
-      .toString()
-      .replace(/(function \(\)\{ return |\(|\)|\; \})/g, '')
-      .trim() === txt[0]
-  ) {
-    return new Watchable(watch[0][0], watch[0][1], () => radiInstance);
+  ll(fn, watch, c) {
+    return watch ? new Link(this, fn, watch, c.split(',')).init() : fn;
   }
-
-  const len = watch.length;
-
-  args.s = fn.call(radiInstance);
-  args.a = new Array(len);
-  args.t = new Array(len);
-  args.__path = `$link-${linkNum}`;
-  linkNum++;
-
-  for (let i = 0; i < len; i++) {
-    const radiInstance = watch[i][0];
-    const field = watch[i][1];
-    args.a[i] = radiInstance[field];
-    args.t[i] = `$rdi[${i}]`;
-    args.f = args.f.replace(txt[i], args.t[i]);
-
-    const path = `${radiInstance.__path}.${field}`;
-    const p = `${args.__path}.s`;
-
-    radiInstance.$eventService.on(path, (path, value) => {
-      args.a[i] = value;
-      const cache = args.f.call(radiInstance, args.a);
-
-      if (args.s !== cache) {
-        args.s = cache;
-        radiInstance.$eventService.emit(p, args.s);
-      }
-    });
-  }
-
-  args.f = new Function('$rdi', 'return ' + args.f + '();')
-
-  if (len <= 0) return args.s;
-  return new Watchable(args, 's', () => radiInstance);
-};
-
-export function cond(a, e) {
-  return new Condition(a, e, this);
 }
-
-export const list = (data, act) => {
-  if (!data) return '';
-  const SELF = this;
-
-  let link;
-  const fragment = document.createDocumentFragment();
-  const toplink = EMPTY_NODE.cloneNode();
-
-  fragment.appendChild(toplink);
-
-  const cache = data.source[data.prop] || [];
-  const cacheLen = cache.length || 0;
-
-  if (Array.isArray(cache)) {
-    for (let i = 0; i < cacheLen; i++) {
-      fragment.appendChild(act.call(SELF, cache[i], i));
-    }
-  } else {
-    let i = 0;
-    for (const key in cache) {
-      fragment.appendChild(act.call(SELF, cache[key], key, i));
-      i++;
-    }
-  }
-
-  link = fragment.lastChild;
-
-  const w = (a, b) => {
-    if (a === 0) return;
-    if (a > 0) {
-      const len = b.length;
-      const start = len - a;
-      for (let i = start; i < len; i++) {
-        fragment.appendChild(act.call(SELF, b[i], i));
-      }
-      const temp = fragment.lastChild;
-      link.parentElement.insertBefore(fragment, link.nextSibling);
-      link = temp;
-      return;
-    }
-    for (let i = 0; i < Math.abs(a); i++) {
-      const templink = link.previousSibling;
-      link.parentElement.removeChild(link);
-      link = templink;
-    }
-  };
-
-  if (cache.__path) {
-    let len = cacheLen;
-    SELF.$eventService.on(cache.__path, (e, v) => {
-      w(v.length - len, v);
-      len = v.length;
-    });
-  }
-
-  return fragment;
-};
