+++ src/__tests__/Link.test.js
@@ -1,0 +1,8 @@
+import Link from '../Link';
+
+// TODO: Test properly
+describe('Link.js', () => {
+  it('doesn\'t crash', () => {
+    new Link({}, () => {}, {}, 'foobar');
+  });
+});
